set the initial test with service starting up

+++ deploy/script/test_cloudlab.sh
@@ -14,6 +14,7 @@ HOST_CONFIG_PATH=$8
 TASK_CONFIG_PATH=$9
 STATIC_CONFIG_PATH=${10}
 NUM_REQUESTS=${11}
+RUN_EXPERIMENT=${12}
 
 parallel-ssh -h deploy/resources/host_addresses/cloud_lab/test_host  -i "pkill -f dodoor"
 parallel-ssh -h deploy/resources/host_addresses/cloud_lab/test_host  -i "pkill -f stress"
@@ -24,12 +25,11 @@ parallel-ssh -t 0 -h deploy/resources/host_addresses/cloud_lab/test_nodes -i "no
 
 parallel-ssh -t 0 -h deploy/resources/host_addresses/cloud_lab/test_scheduler -i "nohup java -cp dodoor/target/dodoor-1.0-SNAPSHOT.jar edu.cam.dodoor.ServiceDaemon -c ${STATIC_CONFIG_PATH} -hc ${HOST_CONFIG_PATH} -tc ${TASK_CONFIG_PATH} -d true -s true -n false  1>${SCHEDULER_TYPE}_scheduler_service.out 2>/dev/null &"
 
-sleep 20
-
-parallel-ssh -t 0 -h deploy/resources/host_addresses/cloud_lab/test_scheduler -i "nohup java -cp dodoor/target/dodoor-1.0-SNAPSHOT.jar edu.cam.dodoor.client.TaskTracePlayer -c ${STATIC_CONFIG_PATH} -hc ${HOST_CONFIG_PATH} -f dodoor/$DATA_PATH  1>${SCHEDULER_TYPE}_replay.out 2>/dev/null &"
-
-# Wait for the tasks to complete
-parallel-ssh -h deploy/resources/host_addresses/cloud_lab/test_scheduler -i "export PYTHONPATH=$HOME/Code/scheduling/dodoor && python3 wait_for_task_completion.py --log scheduler_metrics.log --num_requests ${NUM_REQUESTS}"
-
-parallel-ssh -h deploy/resources/host_addresses/cloud_lab/test_host  -i "pkill -f dodoor"
-parallel-ssh -h deploy/resources/host_addresses/cloud_lab/test_host  -i "pkill -f stress"
+if [ "$RUN_EXPERIMENT" = "true" ]; then
+  sleep 20
+  parallel-ssh -t 0 -h deploy/resources/host_addresses/cloud_lab/test_scheduler -i "nohup java -cp dodoor/target/dodoor-1.0-SNAPSHOT.jar edu.cam.dodoor.client.TaskTracePlayer -c ${STATIC_CONFIG_PATH} -hc ${HOST_CONFIG_PATH} -f dodoor/$DATA_PATH 1>${SCHEDULER_TYPE}_replay.out 2>/dev/null &"
+  # Wait for the tasks to complete
+  parallel-ssh -h deploy/resources/host_addresses/cloud_lab/test_scheduler -i "export PYTHONPATH=$HOME/Code/scheduling/dodoor && python3 wait_for_task_completion.py --log scheduler_metrics.log --num_requests ${NUM_REQUESTS}"
+  parallel-ssh -h deploy/resources/host_addresses/cloud_lab/test_host  -i "pkill -f dodoor"
+  parallel-ssh -h deploy/resources/host_addresses/cloud_lab/test_host  -i "pkill -f stress"
+fi
